ml-service/requirements.txt
# BirdSight ML Service — Python dependencies
torch
torchvision
fastapi
uvicorn[standard]
Pillow
python-multipart
